Added quadtree hover support to volcano. Did some generalizing and refactoring of shared componenets between manhattan/grin2/volcano
